# task-manager-api/dist/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Task Manager",
        "version": "1.0.0",
        "description": "API for managing tasks with Google OAuth"
    },
    "servers": [
        { "url": "https://task-manager-api-9tji.onrender.com" },
        { "url": "http://localhost:5000" }
    ],
...
    "components": {
        "securitySchemes": {
            "OAuth2": {
                "type": "oauth2",
                "flows": {
                    "authorizationCode": {
                        "authorizationUrl": "/auth/google",
                        "tokenUrl": "https://oauth2.googleapis.com/token",
                        "scopes": {
                            "profile": "Access user profile",
                            "email": "Access user email"
                        }
                    }
                }
            }
        },
        "schemas": {
...
            "User": {
                "type": "object",
                "required": ["username", "email"],
                "properties": {
                    "_id": { "type": "string" },
                    "username": { "type": "string" },
                    "email": { "type": "string" },
                    "password": { "type": "string" },
                    "googleId": { "type": "string" },
                    "createdAt": { "type": "string", "format": "date-time" }
                },
                "example": {
                    "username": "Kim Miner",
                    "email": "kimah72@gmail.com",
                    "password": "securepass123"
                }
            }
        }
    },
    "paths": {
...
        "/auth/profile": {
            "get": {
                "tags": ["Auth"],
                "summary": "Get authenticated user profile",
                "security": [{ "OAuth2": ["profile", "email"] }],
                "responses": {
                    "200": {
                        "description": "User profile",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#/components/schemas/User" }
                            }
                        }
                    },
                    "401": { "description": "Not authenticated" }
                }
            }
        },
        "/auth/logout": {
            "get": {
                "tags": ["Auth"],
                "summary": "Log out user",
                "security": [{ "OAuth2": ["profile", "email"] }],
                "responses": {
                    "200": {
                        "description": "Logged out",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": { "type": "string" }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}